api test refacto: api-client.js
